+++ apps/openmw/mwmechanics/alchemy.hpp
@@ -135,7 +135,7 @@ namespace MWMechanics
         /// adjust the skills of the alchemist accordingly.
         /// \param name must not be an empty string, or Result_NoName is returned
 
-        static std::vector<std::string> effectsDescription(const MWWorld::ConstPtr& ptr, const int alchemySKill);
+        static std::vector<std::string> effectsDescription(const MWWorld::ConstPtr& ptr, const float alchemySKill);
     };
 }
